Update the readme with a new journal entry.
I've been doing a lot of cleanup to the codebase, and have made quite a few updates and improvements to the continuous integration/deploy pipeline for the project.  There are now no code smells, no code duplication, and 100% code coverage.  I really want to see this project grow, as I feel this is truly a superior way to log (and the performance is stupid good too).  Almost two years later and I still need to hit the forums to do some shameless plugging...